*Dereference* `SignaturePad` type to be able to use it in the code
There was a compilation error, it didn't find the `SignaturePad` type.
That type was an external library, unable to install it with `npm`, it
had to be installed and imported manually, but the TypeScript compiler
didn't find a reference to it.

Thanks to
https://weblog.west-wind.com/posts/2016/Sep/12/External-JavaScript-dependencies-in-Typescript-and-Angular-2
I found that adding

    declare var SignaturePad: any;

was enough to tell the compiler to *ignore* this type.

It works, but it'll be better to find a way to install SignaturePad with
`npm`.