Import `SignaturePad` as a TypeScript module
After googling about how to import a regular JavaScript library in a
TypeScript project, and lots of wrong articles found, the answer was on
the official [TypeScript
documentation](https://www.typescriptlang.org/docs/handbook/modules.html)

In a `.d.ts` file, the module must be declared (and of course, the
library source code must be available, if the library is installed as a
`npm` module, it helps).

    declare module "signature_pad";

Ok, the module's name matches the folder under `node_modules`, so it may
work by chance.

Then, the library can be imported as

    import * as SignaturePad from 'signature_pad';

It doesn't work as:

    improt SignaturePad from 'signature_pad';

and I don't know why.